improve robustness against callback errors
this mainly relates to failing return value type conversions. Throwing
a Lua error causes the loop to exit in an unrecoverable way. Instead,
we print the errors to stderr and return an appropriate dbus-error to
the client.

This also fixes up several error paths in msg_fromlua that could
throw.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>