bus_gc: only cleanup loop if non-default bus
the event loop was unconditionally cleaned up when collecting a bus,
even when this was a default bus for which more user may exists. This
resulted in :exit_loop() complaining about the bus not being attached.

Correct this by only cleaning up the loop during GC of non-default bus
objects.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>